reorganizing code PROCERGS#707
