+++ src/LoginCidadao/DynamicFormBundle/DynamicFormEvents.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace LoginCidadao\CoreBundle;
+namespace LoginCidadao\DynamicFormBundle;
 
 class DynamicFormEvents
 {
