+++ src/LoginCidadao/DynamicFormBundle/Form/DynamicPersonType.php
@@ -1,6 +1,14 @@
 <?php
+/**
+ * This file is part of the login-cidadao project or it's bundles.
+ *
+ * (c) Guilherme Donato <guilhermednt on github>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
-namespace LoginCidadao\CoreBundle\Form\Type\DynamicForm;
+namespace LoginCidadao\DynamicFormBundle\Form;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Form\FormBuilderInterface;
