+++ src/LoginCidadao/DynamicFormBundle/Service/DynamicFormService.php
@@ -16,7 +16,7 @@ use FOS\UserBundle\Event\FormEvent;
 use FOS\UserBundle\Event\GetResponseUserEvent;
 use FOS\UserBundle\FOSUserEvents;
 use FOS\UserBundle\Model\UserManagerInterface;
-use LoginCidadao\CoreBundle\DynamicFormEvents;
+use LoginCidadao\DynamicFormBundle\DynamicFormEvents;
 use LoginCidadao\CoreBundle\Entity\City;
 use LoginCidadao\CoreBundle\Entity\Country;
 use LoginCidadao\CoreBundle\Entity\PersonAddress;
